uucore/buf_copy: split into platform-specific files
refactor the buf_copy module into Linux and non-Linux implementations.
