+++ src/uucore/src/lib/features/buf_copy/other.rs
@@ -1,0 +1,25 @@
+use std::io::{Read, Write};
+
+use crate::error::UResult;
+
+/// Copy data from `Read` implementor `source` into a `Write` implementor
+/// `dest`. This works by reading a chunk of data from `source` and writing the
+/// data to `dest` in a loop, using std::io::copy. This is implemented for
+/// non-Linux platforms.
+///
+/// # Arguments
+/// * `source` - `Read` implementor to copy data from.
+/// * `dest` - `Write` implementor to copy data to.
+///
+/// # Returns
+///
+/// Result of operation and bytes successfully written (as a `u64`) when
+/// operation is successful.
+pub fn copy_stream<R, S>(src: &mut R, dest: &mut S) -> UResult<u64>
+where
+    R: Read,
+    S: Write,
+{
+    let result = std::io::copy(src, dest)?;
+    Ok(result)
+}
